+++ epic_app/serializers/__init__.py
@@ -17,3 +17,8 @@ from epic_app.serializers.question_serializer import (
     QuestionSerializer,
 )
 from epic_app.serializers.report_serializer import ProgramReportSerializer
+from epic_app.serializers.summary_serializer import (
+    SummaryEvolutionSerializer,
+    SummaryLinkagesSerializer,
+    SummaryOrganizationEvolutionSerializer,
+)