+++ docs/changelog.md
@@ -1,3 +1,48 @@
+## v1.5.1 (2022-09-09)
+
+### Fix
+
+- **eram_visuals_wrapper.py**: Rename file only if it actually exists (duh)
+- allow 'missing_ok' to silence filenotfounderror while unlinking files
+- **eram_visuals_wrapper.py**: Replacement of backup when required was not happening correctly
+- Replaced the use of the statistics mean library as its giving issues in the server
+- Made the mean calculation based on explicit lists
+
+## v1.5.0 (2022-09-09)
+
+### Feat
+
+- Exposed the media folder so the generated evolution summary resources can be reached
+
+## v1.4.1 (2022-09-08)
+
+### Fix
+
+- Removed wrong import
+
+## v1.4.0 (2022-09-08)
+
+### Feat
+
+- **summary_serializer.py**: Added logic to execute the r snippet and tests to validate its correct functioning
+- Corrected summaryEvolutionSerializer as the average should be average of program evo questions for a user, then for its organization, then for the whole set of organizations
+- Added exporter for serialized evolution data into CSV
+
+## v1.3.0 (2022-09-07)
+
+### Feat
+
+- We now import multiple references for a program
+
+## v1.2.0 (2022-09-07)
+
+### Feat
+
+- Created endpoint for evolution-graph, which will call to the R script and generate a new evolution graph for all organizations
+- **views.py**: It is now no longer needed to provide a useless 'pk' when requesting summaries for either linkages or evolution
+- **summary_serializer.py**: Added logic to summarize evolution questions
+- Added new endpoint for generating linkages summary in a json format
+
 ## v1.1.0 (2022-07-01)
 
 ### Feat